backend/requirements.txt
# Django and DRF
Django>=4.2.0,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0
django-environ>=0.11.0
django-extensions>=3.2.0

# Database
psycopg2-binary>=2.9.0
pymongo>=4.6.0

# Authentication & Security
djangorestframework-simplejwt>=5.3.0
cryptography>=41.0.0
bcrypt>=4.1.0
pyotp>=2.9.0
qrcode[pil]>=7.4.2

# Background Tasks
celery>=5.3.0
redis>=5.0.0
django-celery-beat>=2.5.0

# API Documentation
drf-spectacular>=0.26.0

# File Storage
boto3>=1.34.0
django-storages>=1.14.0
Pillow>=10.0.0

# Payments
stripe>=7.8.0

# ML Integration (will be in separate ML requirements)
pandas>=2.1.0
numpy>=1.25.0
scikit-learn>=1.3.0

# Monitoring & Logging
sentry-sdk>=1.38.0

# Development
pytest>=7.4.0
pytest-django>=4.7.0
pytest-cov>=4.1.0
factory-boy>=3.3.0
black>=23.11.0
flake8>=6.1.0
isort>=5.12.0

# Utilities
python-decouple>=3.8
requests>=2.31.0
python-dateutil>=2.8.0
pytz>=2023.3
